Update paper.md
Line 208. Last sentence. "These two where the syntethic data description cases where we focused on during the BioHackathon 2022" --> "These two were the synthetic data description cases which we focused on during the BioHackathon 2022"
Bioschemas section: typo syntethic fixed and changed to synthetic.
Line 236. Second sentence. "We published the version 1 of the metadata model which is publicly available for open discussion and review on GitHub for the community" --> We published the version 1 of the metadata model, available for open discussion and review on GitHub for the community
I suggest adding the link to version 1 in GitHub to that sentence (I do not have it at hand so I did not add it).